AI/requirements.txt
langchain
python-dotenv
langchain-community
pypdf
wikipedia
langchain-text-splitters
lxml
langchain_chroma
sentence_transformers
langchain_huggingface
faiss-cpu
bs4
langchain_ollama
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
arxiv
duckduckgo-search ~=5.3.1b1
psycopg2
verovio
torchvision
tiktoken
accelerate
pydantic
requests
python-multipart